fix: uncaught exception when decoding invalid images
Previously, attempting to decode images using `load_image` failed
without explicitly stating where and why the operation failed.

If the image does not exist or the file does not contain pixel data,

For example, we use `np.frombuffer` to interpret the image data as
an 1-dimensional array. This operation _DOES NOT_ fail when the
image file is empty.

In contrast, passing invalid or empty image data to `cv2.imdecode`
_will_ raise an execption. Which is what happens when we don't
check for a buffer size > 0.

This PR resolves these issues mentioned above by checking
if the computed `chunk_arr` is empty, and if so raising a
legible `ValueError`.

Finally, we assume that if the `cv2.imdecode` fails it's because
the image file's data is corrupt in one way or another and raise
`ValueError` for that too.

Fixes #890